add window update case
